Project_7/deployment/api/requirements.txt
# === API FastAPI ===
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
starlette==0.46.2
# === Interface Streamlit ===
streamlit==1.46.1
requests==2.32.4
# === Machine Learning ===
tensorflow==2.19.0
mlflow==3.1.1
numpy==2.1.3
pandas==2.3.0
# === Traitement de texte ===
nltk==3.9.1
# === Utilitaires ===
python-dotenv==1.1.1
# === Dépendances système ===
protobuf==3.20.3
